refactor: broad codebase improvements across adapters, core, TLS and tests
- Updated inbound/outbound adapters (HTTP, SOCKS, VMess, VLESS, Shadowsocks)
- Enhanced sb-core components (router, DNS, services, transport)
- Improved TLS/Reality configuration and uTLS support
- Added decision documents for Mobile, Tailscale, and TLS
- Updated various integration and e2e tests
- Misc fixes in admin, diagnostics, and platform monitoring

+++ app/src/admin_debug/reloadable.rs
@@ -766,7 +766,7 @@ mod tests_concurrency_and_rollback {
             let _ = apply_with_dryrun(
                 &crate::admin_debug::endpoints::config::ConfigDelta {
                     // Start from 4001 to ensure the first apply changes something
-                    timeout_ms: Some(4001 + i as u64),
+                    timeout_ms: Some(4001 + i),
                     ..Default::default()
                 },
                 false,
